refactor: getx 상태관리 추가
